Utility for deletion of images and invoices

+++ controllers/admin.js
@@ -1,6 +1,7 @@
 const Product = require('../models/product');
 const { validationResult } = require('express-validator/check');
 const mongoose = require('mongoose');
+const fileHelper = require('../util/file');
 
 exports.getAddProduct = (req, res, next) => {
     //send method automatically sets content type to HTML
@@ -28,13 +29,12 @@ exports.postAddProduct = (req, res, next) => {
     const image = req.file;
     const price = req.body.price;
     const description = req.body.description;
-    if(!image)
-    {
+    if (!image) {
         console.log("image error");
         return res.status(422).render('admin/edit-product', {
-            pageTitle: 'Add Product', 
-            path: '/admin/add-product', 
-            editing: false, 
+            pageTitle: 'Add Product',
+            path: '/admin/add-product',
+            editing: false,
             hasError: true,
             product: {
                 title: title,
@@ -47,13 +47,12 @@ exports.postAddProduct = (req, res, next) => {
     }
     const errors = validationResult(req);
 
-    if(!errors.isEmpty())
-    {
+    if (!errors.isEmpty()) {
         console.log("error is not empty")
         return res.status(422).render('admin/edit-product', {
-            pageTitle: 'Add Product', 
-            path: '/admin/add-product', 
-            editing: false, 
+            pageTitle: 'Add Product',
+            path: '/admin/add-product',
+            editing: false,
             hasError: true,
             product: {
                 title: title,
@@ -107,9 +106,9 @@ exports.getEditProduct = (req, res, next) => {
                 return res.redirect('/');
             }
             res.render('admin/edit-product', {
-                pageTitle: 'Edit Product', 
-                path: '/admin/edit-product', 
-                editing: true, 
+                pageTitle: 'Edit Product',
+                path: '/admin/edit-product',
+                editing: true,
                 product: product,
                 hasError: false,
                 errorMessage: null,
@@ -160,12 +159,11 @@ exports.postEditProduct = (req, res, next) => {
 
     const errors = validationResult(req);
 
-    if(!errors.isEmpty())
-    {
+    if (!errors.isEmpty()) {
         return res.status(422).render('admin/edit-product', {
-            pageTitle: 'Edit Product', 
-            path: '/admin/edit-product', 
-            editing: true, 
+            pageTitle: 'Edit Product',
+            path: '/admin/edit-product',
+            editing: true,
             hasError: true,
             product: {
                 title: updatedTitle,
@@ -187,7 +185,8 @@ exports.postEditProduct = (req, res, next) => {
         product.title = updatedTitle;
         product.price = updatedPrice;
         product.description = updatedDesc;
-        if(image) {
+        if (image) {
+            fileHelper.deleteFile(product.imageUrl);
             product.imageUrl = image.path;
         }
         return product
@@ -197,11 +196,11 @@ exports.postEditProduct = (req, res, next) => {
                 console.log('UPDATED PRODUCT');
             })
     })
-    .catch(err => {
-        const error = new Error(err)
-        error.httpStatusCode = 500;
-        return next(error);
-    });
+        .catch(err => {
+            const error = new Error(err)
+            error.httpStatusCode = 500;
+            return next(error);
+        });
 };
 
 exports.getProducts = (req, res, next) => {
@@ -230,7 +229,14 @@ exports.postDeleteProduct = (req, res, next) => {
     const prodId = req.body.productId;
     // Product.deleteById(prodId);
     // Product.findByIdAndRemove(prodId)
-    Product.deleteOne({_id: prodId, userId: req.user._id})
+    Product.findById(prodId)
+        .then(product => {
+            if (!product) {
+                return next(new Error('Product not found!'))
+            }
+            fileHelper.deleteFile(product.imageUrl);
+            return Product.deleteOne({ _id: prodId, userId: req.user._id });
+        })
         .then(() => {
             res.redirect('/admin/products');
             console.log("DESTROYED PRODUCT");
